% Generated by roxygen2: do not edit by hand
\name{ojp_experts}
\alias{ojp_experts}
\title{Fetch PSRAC risk assessment experts}
\usage{
ojp_experts(state = NULL)
}
\arguments{
\item{state}{Optional state filter}
}
\value{
tibble with expert columns
}
\description{
Fetch PSRAC risk assessment experts
}
