% Generated by roxygen2: do not edit by hand
\name{ojp_risk_landscape}
\alias{ojp_risk_landscape}
\title{Fetch PSRAC risk assessment landscape by state}
\usage{
ojp_risk_landscape()
}
\value{
tibble: abbreviation, state, region, jail, parole, pretrial, prison, probation, release
}
\description{
Fetch PSRAC risk assessment landscape by state
}
