% Generated by roxygen2: do not edit by hand
\name{ojp_search}
\alias{ojp_search}
\title{Search OJP datasets by keyword}
\usage{
ojp_search(query)
}
\arguments{
\item{query}{Character string to search in dataset names and descriptions}
}
\value{
tibble: matching datasets
}
\description{
Search OJP datasets by keyword
}
